pyproject.toml
[project]
name = "gh-stats"
version = "0.0.1"
description = "Compiler Explorer GitHub stats"
authors = [
    {name = "Matt Godbolt", email = "matt@godbolt.org"},
    {name = "partouf", email = "partouf@gmail.com"}
]
requires-python = ">=3.9"
dependencies = [
    "pygithub>=1.57",
    "click>=8.1.3",
    "matplotlib>=3.7.0",
]

[tool.black]
line-length = 120
target-version = ['py310']
